feat: add auth feature
