+++ src/config/api-endpoints.config.ts
@@ -1,10 +1,4 @@
 export const ApiEndpoints = {
-  Signup: '/auth/signup',
-  Signin: '/auth/signin',
-  Tokens: '/auth/tokens',
-  Logout: '/auth/logout',
-  Google: '/auth/google',
-
   User: '/user',
   UserMe: '/user/me',
   UserTheme: '/user/theme',
