+++ src/features/auth/auth.types.ts
@@ -1,0 +1,9 @@
+import type { User } from 'types'
+
+export type AuthResponse = {
+  refreshToken: string
+  accessToken: string
+  user: User
+}
+
+export type Tokens = Omit<AuthResponse, 'user'>
