+++ src/features/auth/components/index.ts
@@ -1,4 +1,5 @@
 export * from './AuthFormNavigation'
 export * from './AuthLayout'
+export * from './GoogleSignin'
 export * from './SigninForm'
 export * from './SignupForm'
