+++ src/features/auth/config/auth-api-endpoints.config.ts
@@ -1,0 +1,7 @@
+export enum AuthApiEndpoints {
+  Signup = '/auth/signup',
+  Signin = '/auth/signin',
+  Tokens = '/auth/tokens',
+  Logout = '/auth/logout',
+  Google = '/auth/google'
+}
