+++ src/features/auth/config/auth-cache-keys.config.ts
@@ -1,0 +1,5 @@
+export enum AuthCacheKeys {
+  Signup = 'signup',
+  Signin = 'signin',
+  Logout = 'logout'
+}
