+++ src/pages/HomePage.tsx
@@ -1,6 +1,5 @@
+import { GoogleSignin } from 'features/auth/components/GoogleSignin'
 import { Link } from 'react-router-dom'
-
-import { GoogleSignin } from 'components/auth/GoogleSignin'
 
 import { Pages } from 'config'
