+++ src/redux/user.slice.ts
@@ -1,5 +1,6 @@
 import type { PayloadAction } from '@reduxjs/toolkit'
-import type { AuthResponse, Tokens, User } from 'types'
+import type { AuthResponse, Tokens } from 'features/auth/auth.types'
+import type { User } from 'types'
 
 import { createSlice } from '@reduxjs/toolkit'
