+++ src/types/index.ts
@@ -1,2 +1,2 @@
-export * from './auth.types'
 export * from './board.types'
+export * from './user.types'
